Upgrade SvelteKit with major design overhaul
Remove /posts/posts.json API and use +page.server.js which explicitly only runs
on the server side, which for a static site is at build time. This does not bundle
the posts which are imported into the JS files, and instead only bundles a data JSON
file with each path. This is much cleaner and simpler than before.

For other changes related to the upgrade, check this GitHub issue:
https://github.com/sveltejs/kit/discussions/5774